Refactoring and documenting. Adding templates to data transport for messages
